organize files to folder of entity and repository
