+++ CoinConverter/src/main/kotlin/Exchange.kt
@@ -1,0 +1,4 @@
+import java.math.BigDecimal
+
+data class Exchange(val coin_src: String, val coin_dest: String, val value_src: BigDecimal, val value_dest: BigDecimal, val tax:Float) {
+}
